rename NamedMobInvulnerability to ProtectNamedMobs, let containertrusted players attack the mob
