+++ src/main/java/me/ryanhamshire/GPFlags/FlagsDataStore.java
@@ -268,8 +268,8 @@ public class FlagsDataStore {
 		this.addDefault(defaults, Messages.DisabledRaidMemberOnly, "Anyone can trigger raids in this area", null);
 		this.addDefault(defaults, Messages.RaidMemberOnlyDeny, "You can not initiate a raid in this area", null);
 
-		this.addDefault(defaults, Messages.EnabledNamedMobInvulnerability, "Named mobs will no longer take damage in this area", null);
-		this.addDefault(defaults, Messages.DisabledNamedMobInvulnerability, "Named mobs will continue to take damage in this area", null);
+		this.addDefault(defaults, Messages.EnabledProtectNamedMobs, "Named mobs will no longer take damage in this area", null);
+		this.addDefault(defaults, Messages.DisabledProtectNamedMobs, "Named mobs will continue to take damage in this area", null);
 
 		//load the config file
 		FileConfiguration config = YamlConfiguration.loadConfiguration(new File(messagesFilePath));
