--- src/main/java/me/ryanhamshire/GPFlags/flags/FlagDef_NamedMobInvulnerability.java
@@ -1,51 +1,0 @@
-package me.ryanhamshire.GPFlags.flags;
-
-import me.ryanhamshire.GPFlags.*;
-import org.bukkit.entity.Entity;
-import org.bukkit.entity.EntityType;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
-import org.bukkit.event.entity.EntityDamageEvent;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class FlagDef_NamedMobInvulnerability extends FlagDefinition {
-
-    @EventHandler(priority = EventPriority.LOWEST)
-    public void onEntityDamage(EntityDamageEvent event) {
-        Entity entity = event.getEntity();
-
-        Flag flag = this.GetFlagInstanceAtLocation(entity.getLocation(), null);
-        if (flag == null) return;
-        if (entity.getType() == EntityType.PLAYER) return;
-        if (entity.getCustomName() == null) return;
-
-        event.setCancelled(true);
-    }
-
-    public FlagDef_NamedMobInvulnerability(FlagManager manager, GPFlags plugin) {
-        super(manager, plugin);
-    }
-
-    @Override
-    public String getName() {
-        return "NamedMobInvulnerability";
-    }
-
-    @Override
-    public MessageSpecifier getSetMessage(String parameters) {
-        return new MessageSpecifier(Messages.EnabledNamedMobInvulnerability);
-    }
-
-    @Override
-    public MessageSpecifier getUnSetMessage() {
-        return new MessageSpecifier(Messages.DisabledNamedMobInvulnerability);
-    }
-
-    @Override
-    public List<FlagType> getFlagType() {
-        return Arrays.asList(FlagType.CLAIM, FlagType.WORLD, FlagType.SERVER);
-    }
-
-}
